Remove conda-verify from requirements as it is not longer maintained or used it seems Adjust regex to allow for parsing additional ld version strings Only run models in model repo if the required language is installed Update serialization utilities for use with pandas Added option to pandas serializer for specifying the index column/row Check for type transformations in code generation, discarding statement/function transforms Allow array to be passed to type2numpy and definition2dtype to fill in missing information in types like precision for flexible precision types (bytes, unicode, string) Allow timesync as model from input Update existing repo on run Added support to table serializer for getting number of columns from first message for datatypes that only specify a type for all columns Fix bug where allowWrapped was causing default data type to be treated as non-default Add option for setup_test_env utility to install the version of yggdrasil on conda-forge/PyPI Add specialized errors for build process Check for scope in target_link_libraries via regex Change default build directory Allow split server in other i/o channel signatures Updated rapidjson & python-rapidjson submodules

+++ tests/test_services.py
@@ -251,15 +251,17 @@ def test_validate_model_repo():
     import tempfile
     dest = os.path.join(tempfile.gettempdir(), "model_repo")
     url = "https://github.com/cropsinsilico/yggdrasil_models"
+    dest_exists = os.path.isdir(dest)
     for x in [url, url + "_test"]:
         try:
-            repo = git.Repo.clone_from(x, dest)
+            repo = None
+            if not dest_exists:
+                repo = git.Repo.clone_from(x, dest)
             model_dir = os.path.join(dest, "models")
-            if os.path.isdir(model_dir):
-                # This condition can be removed once there are models
-                # in the non-dev repository
-                validate_model_submission(model_dir)
-            repo.close()
+            validate_model_submission(model_dir)
+            if not dest_exists:
+                repo.close()
         finally:
-            if os.path.isdir(dest):
-                git.rmtree(dest)
+            if not dest_exists:
+                if os.path.isdir(dest):
+                    git.rmtree(dest)
